LayoutTests/fast/images/text-recognition/image-overlay-do-not-translate-expected.txt

PASS imageOverlayContainer(image1).translate is false
PASS imageOverlayContainer(image2).translate is false
PASS successfullyParsed is true

TEST COMPLETE
